Updates when chages occur in db

+++ KreemMachine/KreemMachineLibrary/Services/UserService.cs
@@ -1,5 +1,6 @@
 ﻿using KreemMachineLibrary.Helpers;
 using KreemMachineLibrary.Models;
+using KreemMachineLibrary.Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -7,6 +8,9 @@ using System.Linq;
 using System.Data.Entity;
 using System.Text;
 using System.Threading.Tasks;
+using System.Data.SqlClient;
+using System.Windows.Forms;
+using System.Windows;
 
 namespace KreemMachineLibrary.Services
 {
@@ -14,14 +18,16 @@ namespace KreemMachineLibrary.Services
     {
         
         DataBaseContext db = Globals.db;
-
+        
+        
+ 
         public UserService()
         {
             // Load all the roles into memory, 
             // since it's just a few of them that are probable to be used anyway
             // it's better to load them in memory in advance
             // than to have a sepparate query when we actually need them
-           // db.Roles.Load();
+            // db.Roles.Load();
         }
 
         public User Save(User user)
@@ -47,6 +53,7 @@ namespace KreemMachineLibrary.Services
 
         internal User SaveToDatabase (User user)
         {
+
             var fromDb = db.Users.Add(user);
             db.SaveChanges();
             
@@ -86,9 +93,26 @@ namespace KreemMachineLibrary.Services
             db.SaveChanges();
         }
 
-        public void DeleteEmployee(User user) {
-            db.Users.Remove(user);
-            db.SaveChanges();
+        public int DeleteEmployee(User user, User logedUser) {
+            Role[] notDeletableEmployeeRoles = new Role[] { Role.Administrator, Role.Manager };
+            if (user == logedUser) {
+                DialogResult result = MessageBox.Show("Are you sure you want to delet you own accout? You will be logged out!", MessageBoxButtons.YesNoCancel.ToString());
+                if(result == DialogResult.OK)
+                {
+                    db.Users.Remove(user);
+                    db.SaveChanges();
+                    return -1;
+                }
+                
+            }
+            else if (notDeletableEmployeeRoles.Contains(user.Role)) {
+                throw new DeletAdminAccountException("You are not allow to delet an employee with power");
+            } else {
+                db.Users.Remove(user);
+                db.SaveChanges();               
+            }
+            return 1;
+
         }
 
         public string GenerateEmployeeEmail(string first, string last) {
